Update class state of ui for good visualisation
